+++ src/main/java/com/braintreegateway/PaymentInstrumentType.java
@@ -6,4 +6,5 @@ public final class PaymentInstrumentType {
     public static final String PAYPAL_ACCOUNT= "paypal_account";
     public static final String SEPA_BANK_ACCOUNT = "sepa_bank_account";
     public static final String CREDIT_CARD = "credit_card";
+    public static final String APPLE_PAY_CARD = "apple_pay_card";
 }