+++ src/main/java/com/braintreegateway/test/Nonce.java
@@ -5,4 +5,8 @@ public class Nonce {
     public static String Consumed = "fake-consumed-nonce";
     public static String PayPalOneTimePayment = "fake-paypal-one-time-nonce";
     public static String PayPalFuturePayment = "fake-paypal-future-nonce";
+    public static String ApplePayVisa = "fake-apple-pay-visa-nonce";
+    public static String ApplePayAmex = "fake-apple-pay-amex-nonce";
+    public static String ApplePayMastercard = "fake-apple-pay-mastercard-nonce";
+    public static String AbstractTransactable = "fake-abstract-transactable-nonce";
 }