+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 2.27.0
+
+* Add risk_data to Transaction and Verification with Kount decision and id
+* Add verification_amount an option when creating a credit card
+* Add TravelCruise industry type to Transaction
+* Add room_rate to Lodging industry type
+* Add CreditCard#verification as the latest verification on that credit card
+* Add ApplePay support to all endpoints that may return ApplePayCard objects
+* Add prefix to sample Webhook to simulate webhook query params
+
 ## 2.36.0
 
 * ApplePay support